moved the contact form js file to components folder
